% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoPredictR_functions.R
\name{PseudoLayer}
\alias{PseudoLayer}
\title{A nastyverse function}
\usage{
PseudoLayer(inputRas, pseudoRas = "fathorse", mask = F)
}
\arguments{
\item{inputRas}{Your prediction layer}

\item{pseudoRas}{character string specifiying which built-in classical painting to use as pseudo predictor. Defaults to fathorse.}
}
\description{
This function allows you play some nasty beats to accompany your code.
}
\details{
use pseudoRas="show me" to see all possible arguments for the pseudoRas param
}
\examples{
# Get a fathorse raster that matches the extent, and resolution of your prediction layers 
beatr("tubular")
# Update all packages and hear Good Morning Vietnam! when it is ready
update.packages(ask=FALSE); beatr("ecoroms")
}
\keyword{dank}
